Input box will not be disabled if not close enough to the school

+++ app/src/main/java/heatchat/unite/com/heatchat/models/ChatMessage.java
@@ -27,7 +27,7 @@ public class ChatMessage {
         this.lat = lat;
 
         // Initialize to current time
-        this.time = (long) (new Date().getTime() / 1000);
+        this.time = new Date().getTime() / 1000;
     }
 
     public ChatMessage(String uid, String text, double lat, double lon, long time) {
